fix bug

+++ app/Http/Controllers/Resource/TransactionFieldController.php
@@ -7,8 +7,15 @@ use App\Models\TransactionField;
 
 class TransactionFieldController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware('isFeatureAccessible:Pad Management');
+    }
+
     public function destroy(TransactionField $transactionField)
     {
+        $this->authorize('delete', $transactionField->transaction);
+
         TransactionField::where('line', $transactionField->line)->forceDelete();
 
         return back()->with('successMessage', 'Deleted Successfully');
